Move PetService-related classes to another package and add a Component Scan to the main entry class file.
