+++ src/main/java/com/springframework/pets/PetService.java
@@ -1,4 +1,4 @@
-package student.springframework.sfsdi.services;
+package com.springframework.pets;
 
 public interface PetService {
     String getPetType();
